requirements.txt
discord.py==2.3.2
aiodns==3.1.1
tinydb==4.8.0